server/requirements.txt
Flask==2.3.3
paho-mqtt==1.6.1
Flask-PyMongo==2.3.0
pymongo==4.5.0
Werkzeug==2.3.7
